Track and update notes rather than always creating.
These notes are tracked based on their title, which should always
be unique as it is based on date, time, your account and the other
person's account.

This also allows log files to stay on disk. In fact, appropriate
Pidgin logs require this. Pidgin will create the log file on the
first message submission and then updates that log file as the
conversation progresses. If that log file isn't there, you get
an incomplete log containing whatever was logged before the file
was removed.