Conditionally update note based on whether log changed.
This will avoid updating logs that haven't changed since the last time
the log was polled.

More complicated determinations of old logs will be needed to avoid
even bothering with logs that are several days, months or years old.
This change does not accomplish that yet.

There is a cost to downloading notes as well that could be viewed as
similar to the cost of simply updating the note. That said, this avoids
arbitrarily increasing the "note updated" time on the note despite
there being no changes.